requirements.txt
requests
google-genai
apscheduler
python-dotenv
groq
